move grid content  into base class because it might be useful for testing anything related with shader.
