--- osu.Framework.Font.Tests/Visual/VisualTestSample.cs
@@ -1,13 +1,0 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using NUnit.Framework;
-using osu.Framework.Testing;
-
-namespace osu.Framework.Font.Tests.Visual
-{
-    [TestFixture]
-    public sealed class VisualTestSample : TestScene
-    {
-    }
-}
